% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSistau.R
\name{getPax}
\alias{getPax}
\title{getPax}
\source{
CNRT.
}
\usage{
getPax(annual = TRUE, type = "TODOS")
}
\arguments{
\item{annual}{`TRUE` or `FALSE`. If `TRUE`, returns annual data. If `FALSE` returns monthly data., Default: `TRUE`.}

\item{type}{Choose the transport mode. It can be "COLECTIVO", "TREN", "SUBTE" o "TODOS"}
}
\value{

}
\description{
Get all available pax data by period.
}
\references{

}
\author{
Juan Ignacio Fulponi
}
