% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSistau.R
\name{plotCCP}
\alias{plotCCP}
\alias{plotsubsidiosycomp}
\title{plotCCP}
\source{
Ministerio de Obras Públicas y Ministerio de Transporte de la Nación.

Ministerio de Obras Públicas y Ministerio de Transporte de la Nación.
}
\usage{
plotCCP(annual = TRUE)

plotsubsidiosycomp(annual = TRUE)
}
\arguments{
\item{annual}{`TRUE` or `FALSE`. If `TRUE`, returns annual data. If `FALSE` returns monthly data., Default: `TRUE`.}
}
\value{



}
\description{
Plots CCP data by period.

Plots all subsidies data by period.
}
\references{



}
\author{
Juan Ignacio Fulponi

Juan Ignacio Fulponi
}
