% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSistau.R
\name{plotPax}
\alias{plotPax}
\title{plotPax}
\source{
Ministerio de Obras Públicas y Ministerio de Transporte de la Nación.
}
\usage{
plotPax(annual = TRUE, type = "TODOS")
}
\arguments{
\item{annual}{`TRUE` or `FALSE`. If `TRUE`, returns annual data. If `FALSE` returns monthly data., Default: `TRUE`.}

\item{type}{Choose the transport mode. It can be "COLECTIVO", "TREN", "SUBTE" o "TODOS".}
}
\value{

}
\description{
Plots all available pax data by period.
}
\references{

}
\author{
Juan Ignacio Fulponi
}
