% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSistau.R
\name{plotSistau}
\alias{plotSistau}
\title{plotSistau}
\source{
Ministerio de Obras Públicas y Ministerio de Transporte de la Nación.
}
\usage{
plotSistau(annual = TRUE)
}
\arguments{
\item{annual}{`TRUE` or `FALSE`. If `TRUE`, returns annual data. If `FALSE` returns monthly data., Default: `TRUE`.}
}
\value{

}
\description{
Plots SISTAU data by period.
}
\references{

}
\author{
Juan Ignacio Fulponi
}
